requirements.txt
elasticsearch>=5.5.0
kafka-python>=1.4.3
pymongo>=3.6.1
raven[flask]>=6.7.0
python-etcd>=0.4.5
gevent>=1.2.2
threadpool>=1.3.2
arrow>=0.12.1
fire>=0.1.3
pydash>=4.5.0
requests>=2.18.4
schedule>=0.5.0
redis>=2.10.6
